// src/types/LocationArea.ts
// Generated by https://quicktype.io
export interface LocationArea {
  id: number;
  name: string;
  game_index: number;
  encounter_method_rates: EncounterMethodRate[];
  location: Location;
  names: Name[];
  pokemon_encounters: PokemonEncounter[];
}

export interface EncounterMethodRate {
  encounter_method: Location;
  version_details: EncounterMethodRateVersionDetail[];
}

export interface Location {
  name: string;
  url: string;
}

export interface EncounterMethodRateVersionDetail {
  rate: number;
  version: Location;
}

export interface Name {
  name: string;
  language: Location;
}

export interface PokemonEncounter {
  pokemon: Location;
  version_details: PokemonEncounterVersionDetail[];
}

export interface PokemonEncounterVersionDetail {
  version: Location;
  max_chance: number;
  encounter_details: EncounterDetail[];
}

export interface EncounterDetail {
  min_level: number;
  max_level: number;
  condition_values: any[];
  chance: number;
  method: Location;
}